Add scurve
