+++ src/virtual_maize_field/world_generator/world_description.py
@@ -16,7 +16,7 @@ from virtual_maize_field.world_generator.models import (
 )
 
 AVAILABLE_ILANDS = []
-AVAILABLE_SEGMENTS = ["straight", "curved", "island"]
+AVAILABLE_SEGMENTS = ["straight", "curved", "scurved", "island"]
 
 
 class WorldDescription:
@@ -27,8 +27,12 @@ class WorldDescription:
         row_width: float = 0.75,
         rows_count: int = 6,
         row_segments: list[str] = AVAILABLE_SEGMENTS[:2],
-        row_segment_straight_length_min: int = 1,
-        row_segment_straight_length_max: float = 2.5,
+        row_segment_straight_length_min: int = 0.5,
+        row_segment_straight_length_max: float = 1,
+        row_segment_scurved_offset_min: float = 0.5,
+        row_segment_scurved_offset_max: float = 1.5,
+        row_segment_scurved_length_min: float = 3,
+        row_segment_scurved_length_max: float = 5,
         row_segment_curved_radius_min: float = 3.0,
         row_segment_curved_radius_max: float = 10.0,
         row_segment_curved_arc_measure_min: float = 0.3,
@@ -130,7 +134,7 @@ class WorldDescription:
         current_curve = 0
 
         while current_row_length < self.row_length:
-            # Choose rendom segment
+            # Choose random segment
             segment_name = self.rng.choice(self.row_segments)
 
             if segment_name == "straight":
@@ -144,6 +148,38 @@ class WorldDescription:
                 )
 
                 segment = {"type": "straight", "length": length}
+
+                current_row_length += length
+
+            elif segment_name == "scurved":
+                offset = (
+                    self.rng.random()
+                    * (
+                        self.row_segment_scurved_offset_max
+                        - self.row_segment_scurved_offset_min
+                    )
+                    + self.row_segment_scurved_offset_min
+                )
+                length = (
+                    self.rng.random()
+                    * (
+                        self.row_segment_scurved_length_max
+                        - self.row_segment_scurved_length_min
+                    )
+                    + self.row_segment_scurved_length_min
+                )
+
+                if current_row_length + length > self.row_length:
+                    continue
+
+                curve_dir = self.rng.integers(2)
+
+                segment = {
+                    "type": "scurved",
+                    "offset": offset,
+                    "length": length,
+                    "curve_dir": curve_dir,
+                }
 
                 current_row_length += length
